move properties and methods to PeriodArray and TimedeltaArray mixins
